[tx] extract separate file for localising error messages.
